+++ packages/server/src/api/database/index.js
@@ -1,23 +1,12 @@
 const log = require('@home-gallery/logger')('server.api.database');
 
 const { filterEntriesByQuery } = require('@home-gallery/query')
+const { buildEntriesTextCache } = require('@home-gallery/query')
 
 const { waitReadWatch } = require('./read-database');
 const { cache } = require('./cache-middleware');
 const { sanitizeInt } = require('./sanitize');
-
-function uniqEntries(entries) {
-  // unify media entries
-  const idToEntry = entries.reduce((result, entry) => {
-    if (!result[entry.id]) {
-      result[entry.id] = entry;
-    }
-    return result;
-  }, Object.create({}));
-  const uniqEntries = Object.values(idToEntry);
-  uniqEntries.sort((a, b) => a.date < b.date ? 1 : -1);
-  return uniqEntries;
-}
+const { applyEvents } = require('./apply-events')
 
 const filterDatabase = (entries, query, cb) => query ? filterEntriesByQuery(entries, query, cb) : cb(null, entries)
 
@@ -60,19 +49,42 @@ function databaseApi(eventbus) {
     }
   }
 
+  const clearCaches = () => {
+    databaseCache.clear();
+    entryCache = {};
+    log.trace(`Cleared caches`);
+  }
+
+  eventbus.on('userAction', event => {
+    if (!database.data.length) {
+      log.warn(`Received a user action event without a database. Skip event merging for database`);
+      return
+    }
+    const changedEntries = applyEvents(database.data, [event])
+    if (!changedEntries.length) {
+      log.debug(`Event did not change current database`);
+      return
+    }
+    buildEntriesTextCache(changedEntries)
+    log.debug(`Applied user action event to ${changedEntries.length} database entries`);
+    clearCaches()
+    eventbus.emit('server', {
+      action: 'databaseUpdated'
+    })
+
+  })
+
   return {
     /**
      * @param {string} databaseFilename
      */
-    init: (databaseFilename) => {
-      waitReadWatch(databaseFilename, (err, newDatabase) => {
+    init: (databaseFilename, getEvents) => {
+      waitReadWatch(databaseFilename, getEvents, (err, newDatabase) => {
         if (err) {
           log.error(`Could not read database file ${databaseFilename}: ${err}`);
         } else {
-          newDatabase.data = uniqEntries(newDatabase.data)
           database = newDatabase;
-          databaseCache.clear();
-          entryCache = {};
+          clearCaches()
           eventbus.emit('server', {
             action: 'databaseReloaded'
           })